offline_model_trainer/requirements.txt
numpy
pandas
scikit-learn
tensorflow
requests
pyyaml
jupyter